pyproject.toml
[tool.poetry]
name = "voice translation"
version = "0.1.0"
description = ""
authors = ["Artachtron <florent.vaucher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
huggingface-hub = "^0.23.0"
librosa = "^0.10.2"
faster-whisper = "^1.0.2"
pyaudio = "^0.2.14"
speechrecognition = "^3.10.4"
gradio = "^4.31.0"
gtts = "^2.5.1"
pydantic-settings = "^2.2.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
